deepen: inline CronSecFilingService into discovery-handler
`lib/cron/sec-filing-service.ts` was a 128-LOC static-class wrapper
around one method (`checkForNewFilings`) with one production caller
(`lib/cron/handlers/discovery-handler.ts`). The module's own docblock
admitted the static-class shape was a compat hack; every other method
the class once carried had already been deleted as dead orbital.

Fold the RSS-walk loop into `discovery-handler.ts` as a private
`discoverFilingsFromRss(tickers)` function so all discovery
orchestration — ticker enrichment, RSS walk, backlog dedup, bulk job
creation — lives in one deep module. Discovery's interface stays
`handleDiscovery(payload)`; the private walk sits alongside the other
private helpers (`enrichTickersWithCik`, `createBulkFetchJobs`).

Test seam moved down one level: the three tests that previously mocked
`CronSecFilingService.checkForNewFilings` now mock
`checkTickerForNewFilings` from `lib/sec-edgar/ticker-monitoring`
(the real RSS entry point) and drive the raw `RSSFilingEntry[]`
format the handler reshapes. `tests/security/cron-security.test.ts`
dropped a stale mock whose method names (`discoverNewFilings`,
`processFilingsForUser`, `runSecFilingMonitoring`) didn't even exist
on the real module — drift symptom noted in the deleted file's
docblock.

Result: 20 tests across the three suites pass. One fewer file, one
fewer test seam, no callers left branching on a static class.
